Move classes to their files
